+++ src/features/auth/gnb/ui/index.ts
@@ -1,0 +1,2 @@
+export { default as GNBWrapper } from './gnb-wrapper'
+export { default as AuthSidebar } from './auth-sidebar'
