+++ src/features/auth/logout/api/index.ts
@@ -1,0 +1,1 @@
+export * from './logout.API'
