+++ src/features/auth/logout/api/logout.API.ts
@@ -1,4 +1,4 @@
-import axiosInstance from '@/shared/api/axios-instance'
+import { axiosInstance } from '@/shared/api'
 
 export async function postLogout(): Promise<void> {
   await axiosInstance.post('/auth/logout')
