+++ src/features/auth/logout/model/index.ts
@@ -1,0 +1,1 @@
+export { default as useLogoutMutation } from './use-logout-mutation'
