+++ src/features/auth/logout/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as LogoutModal } from './logout-button'
