+++ src/features/auth/model/index.ts
@@ -1,0 +1,2 @@
+export * from './auth-error-messages'
+export * from './get-auth-error-message'
