+++ src/features/auth/signin/ui/email-signin-modal.tsx
@@ -1,9 +1,8 @@
-import { ModalContent, ModalOverlay, ModalPortal, Modal, ModalTitle } from '@/shared/ui/modal'
-import Text from '@/shared/ui/text/text'
+import { ModalContent, ModalOverlay, ModalPortal, Modal, ModalTitle, Text } from '@/shared/ui'
 
 import SigninForm from './signin-form'
 
-import type { AuthModalType } from '../model/auth-modal.type'
+import type { AuthModalType } from '../../types'
 
 interface Props {
   isOpen: boolean
