+++ src/features/auth/signin/ui/index.ts
@@ -1,0 +1,5 @@
+export { default as EmailSigninModal } from './email-signin-modal'
+export { default as SigninForm } from './signin-form'
+
+export { default as LoginSelectButton } from './login-select-button'
+export { default as LoginSelectModal } from './login-select-modal'
