+++ src/features/auth/signup/ui/index.ts
@@ -1,0 +1,2 @@
+export { default as SignupModal } from './signup-modal'
+export { default as SignupForm } from './signup-form'
