+++ src/features/auth/types/index.ts
@@ -1,0 +1,1 @@
+export * from './auth-modal.type'
