--- src/features/auth/withdraw/index.ts
@@ -1,2 +1,0 @@
-export { default as WithdrawModal } from './ui/withdraw-modal'
-export { default as useWithdrawMutation } from './model/use-withdraw-mutation'
