+++ src/features/auth/withdraw/model/index.ts
@@ -1,0 +1,1 @@
+export { default as useWithdrawMutation } from './use-withdraw-mutation'
