+++ src/features/auth/withdraw/model/use-withdraw-mutation.ts
@@ -1,9 +1,9 @@
 import { useMutation } from '@tanstack/react-query'
 
-import { useUserStore } from '@/entities/user/models/use-user-store'
-import { toast } from '@/shared/ui/toast'
+import { useUserStore } from '@/entities/user'
+import { toast } from '@/shared/ui'
 
-import { deleteWithdraw } from '../api/withdraw.API'
+import { deleteWithdraw } from '../api'
 
 export default function useWithdrawMutation() {
   const { clearUser } = useUserStore()
