+++ src/features/auth/withdraw/ui/index.ts
@@ -1,0 +1,1 @@
+export { default as WithdrawModal } from './withdraw-modal'
